requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
psycopg2-binary==2.9.7
pyTelegramBotAPI==4.12.0
python-dotenv==1.0.0
python-telegram-bot-pagination==0.0.2
requests==2.31.0
urllib3==2.0.4
